Move helper testing functions
